+++ ios/NativeCore/Blake3Dispatch.c
@@ -1,1 +1,0 @@
-#include "../../third_party/blake3/c/blake3_dispatch.c"