+++ ios/NativeCore/XxHash.c
@@ -1,1 +1,0 @@
-#include "../../third_party/xxhash/xxhash.c"